🦄 refactor(util、asset): 货币格式化移至项目总util文件夹
