+++ src/utils/index.ts
@@ -7,3 +7,4 @@ export * from './form';
 export * as storage from './storage';
 export * as http from './http';
 export * as dateComputer from './dateComputer';
+export * from './currencyFormatter';
